requirements.txt
Django==2.0.6
django-filter==2.0.0
djangorestframework==3.8.2
mysqlclient==2.0.3
pkg_resources==0.0.0
pytz==2021.1
